dom/network/tests/test_tcpsocket_legacy.html

<!DOCTYPE HTML>
<html>
<!--
Test of legacy navigator interface for opening TCPSocket/TCPServerSocket.
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 885982</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1084245">Mozilla Bug 1084245</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script>
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv(
    { set: [ ['dom.mozTCPSocket.enabled', true] ] },
    setPerms);

  function setPerms() {
    SpecialPowers.pushPermissions(
      [ { type: 'tcp-socket', allow: true, context: document } ],
      runTest);
  }

  function runTest() {
    // See bug 903830; in e10s mode we never get to find out the localPort if we
    // let it pick a free port by choosing 0.  This is the same port the xpcshell
    // test was using.
    var serverPort = 8085;

    var listeningServer = navigator.mozTCPSocket.listen(serverPort,
                                                        { binaryType: 'arraybuffer' },
                                                        -1);
    listeningServer.onconnect = function(ev) {
      ok(true, "got server connect");
      ev.socket.close()
    }

    var clientSocket = navigator.mozTCPSocket.open('127.0.0.1', serverPort,
                                                   { binaryType: 'arraybuffer' });
    clientSocket.onopen = function() { ok(true, "got client open"); }
    clientSocket.onclose = function() {
      ok(true, "got client close");
      SimpleTest.finish();
    }
  }
</script>
</body>
</html>
